requirements.txt
requests
bs4
openai
serpapi
tqdm
python-dotenv
streamlit